Reorganize things a bit
