+++ package.json
@@ -49,7 +49,7 @@
   },
   "scripts": {
     "test": "npm run lint && npm run flow && jest --coverage --testPathIgnorePatterns __tests__/devenv-e2e/",
-    "test:e2e:dev-env": "jest -c __tests__/devenv-e2e/jest.config.js",
+    "test:e2e:dev-env": "jest -c __tests__/devenv-e2e/jest/jest.config.js",
     "clean": "rimraf dist",
     "prepare": "npm run clean && npm run build",
     "postinstall": "node ./helpers/check-version.js",
@@ -92,6 +92,7 @@
     "@babel/preset-env": "7.20.2",
     "@babel/preset-flow": "7.18.6",
     "@jest/globals": "^29.3.1",
+    "babel-core": "7.0.0-bridge.0",
     "babel-jest": "^29.3.1",
     "babel-plugin-module-resolver": "4.1.0",
     "check-node-version": "^4.2.1",
